requirements.txt
altair==4.2.2
attrs==22.2.0
blinker==1.5
boto3==1.26.25
botocore==1.29.76
cachetools==5.3.0
certifi==2022.12.7
chardet==4.0.0
click==8.1.3
decorator==5.1.1
entrypoints==0.4
exceptiongroup==1.1.1
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
idna==2.10
importlib-metadata==6.1.0
iniconfig==2.0.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pluggy==1.0.0
protobuf==3.20.3
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyarrow==11.0.0
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyrsistent==0.19.3
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.25.1
rich==13.3.2
s3transfer==0.6.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.7
streamlit==1.20.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
urllib3==1.26.14
validators==0.20.0
watchdog==3.0.0
zipp==3.15.0